Remove no-use-before-declare.
This rule is only useful if we allow `var` which we don't via `no-var-keyword` in `tslint:recommended`.

It was also generating a warning but we can remove it and use the default value from `tslint:recommended` which is `false`.

From https://palantir.github.io/tslint/rules/no-use-before-declare/

> Since most modern TypeScript doesn’t use var, this rule is generally discouraged and is kept around for legacy purposes. It is slow to compute, is not enabled in the built-in configuration presets, and should not be used to inform TSLint design decisions.

Before:

```bash
$ ~/stellar/js-stellar-sdk $ npx gulp lint:src
[14:47:17] Using gulpfile ~/stellar/js-stellar-sdk/gulpfile.js
[14:47:17] Starting 'lint:src'...
Warning: The 'no-use-before-declare' rule requires type information.
[14:47:18] Finished 'lint:src' after 679 ms
```

After:
```bash
$ ~/stellar/js-stellar-sdk $ npx gulp lint:src
[14:47:49] Using gulpfile ~/stellar/js-stellar-sdk/gulpfile.js
[14:47:49] Starting 'lint:src'...
[14:47:50] Finished 'lint:src' after 658 ms
```